requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
requests==2.31.0
vaderSentiment==3.3.2
scikit-learn==1.3.0
numpy==1.24.3
python-dotenv==1.0.0